+++ persistent-sqlite/ChangeLog.md
@@ -2,8 +2,10 @@
 
 ## 2.10.2
 
+* Add a new `RawSqlite` type and `withRawSqliteConnInfo` function that allow
+  access to the underlying Sqlite `Connection` type. [#772](https://github.com/yesodweb/persistent/pull/772)
 * Expose the internals of `Connection` in an Internal module, allowing the user
-  to call SQLite functions via the C FFI. [#772](https://github.com/yesodweb/persistent)
+  to call SQLite functions via the C FFI. [#772](https://github.com/yesodweb/persistent/pull/772)
 
 ## 2.10.1